refactor: change the user directory
