+++ backend/main/src/User/UseCases/CreateUserUseCase/CreateUserUseCase.ts
@@ -1,5 +1,5 @@
 import { IUserRepository } from "src/User/Domain/IUserRepository";
-import { User, UserId } from "src/User/Domain/User";
+import { User, UserId } from "src/User/Domain/Entity/User";
 import { Err, IUseCase, Ok, Result } from "src/Shared";
 import {
 	CreateUserUseCaseErrorType,
